requirements.txt
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
pymupdf
python-dotenv
streamlit
fastapi
uvicorn